refactor(pauli-sum): split translation symmetry module
